add: Добавлена полная контейнеризация проекта, все готово к подключению реальных запросов на фронтенде вместо моковых данных

+++ backend/src/main/java/com/leviti/backend/modules/user/service/UserServiceImpl.java
@@ -56,7 +56,7 @@ public class UserServiceImpl implements UserService {
                         .theme("SYSTEM")
                         .language("ru")
                         .currency("RUB")
-                        .monthStartDay(1)
+                        .monthStartDay((short) 1)
                         .emailNotifications(true)
                         .pushNotifications(true)
                         .budgetNotifications(true)
